patched zawarudo

+++ Javascript/ZaWarudo/events/messageCreate.js
@@ -11,8 +11,10 @@ module.exports = {
         }
 
 		if (message.content.toLowerCase().includes('go sleep')) {
-			console.log(`[INFO] asked to log out by ${message.author.tag}`);
-			message.reply('OK, gn :sleeping:').then(() => {setTimeout(() => client.destroy(), 1000);});
+			if (message.member.roles.cache.some((r) => r.name === "The World")) {
+				console.log(`[INFO] asked to log out by ${message.author.tag}`);
+				message.reply('OK, gn :sleeping:').then(() => {setTimeout(() => client.destroy(), 1000);});
+			}
 		}
 	},
 };
